# performance improvements to dae