#4619

Two main changes here. 

* Reduce the number of iterations required to convergence
  * Allow return value removal if the call's parameters were not changed
* Improve individual iteration time 
  * Take more advantage of DAEScanner threading by combing aggregating call to function as they are scanned opposed to tracking the calls a function makes and then combine in another pass.  This does introduce a mutex, but with larger binaries that merging after the scan is a large portion of the time

Crude measurements 
Before:
```
[PassRunner]   running pass: dae-optimizing...                     53.1484 seconds.
```
After:
```
[PassRunner]   running pass: dae-optimizing...                     20.1887 seconds.
```

<details>
<summary>More details</summary>
<details>
<summary>Before</summary>

```
[PassRunner] running passes
[PassRunner]   running pass: duplicate-function-elimination...     14.3721 seconds.
[PassRunner]   running pass: memory-packing...                     0.375367 seconds.
[PassRunner]   running pass: once-reduction...                     0.284163 seconds.
[PassRunner]   running pass: ssa-nomerge...                        8.43984 seconds.
[PassRunner]   running pass: dce...                                2.62311 seconds.
[PassRunner]   running pass: remove-unused-names...                0.436566 seconds.
[PassRunner]   running pass: remove-unused-brs...                  2.46523 seconds.
[PassRunner]   running pass: remove-unused-names...                0.421022 seconds.
[PassRunner]   running pass: optimize-instructions...              2.00798 seconds.
[PassRunner]   running pass: pick-load-signs...                    0.591743 seconds.
[PassRunner]   running pass: precompute...                         1.29118 seconds.
[PassRunner]   running pass: optimize-added-constants-propagate... 18.0451 seconds.
[PassRunner]   running pass: code-pushing...                       0.877306 seconds.
[PassRunner]   running pass: simplify-locals-nostructure...        4.57143 seconds.
[PassRunner]   running pass: vacuum...                             3.66881 seconds.
[PassRunner]   running pass: reorder-locals...                     0.735429 seconds.
[PassRunner]   running pass: remove-unused-brs...                  1.81832 seconds.
[PassRunner]   running pass: coalesce-locals...                    4.37666 seconds.
[PassRunner]   running pass: local-cse...                          2.48903 seconds.
[PassRunner]   running pass: simplify-locals...                    6.08436 seconds.
[PassRunner]   running pass: vacuum...                             3.85254 seconds.
[PassRunner]   running pass: reorder-locals...                     0.612518 seconds.
[PassRunner]   running pass: coalesce-locals...                    3.73631 seconds.
[PassRunner]   running pass: reorder-locals...                     0.594931 seconds.
[PassRunner]   running pass: vacuum...                             4.17217 seconds.
[PassRunner]   running pass: code-folding...                       2.27605 seconds.
[PassRunner]   running pass: merge-blocks...                       1.1134 seconds.
[PassRunner]   running pass: remove-unused-brs...                  1.94706 seconds.
[PassRunner]   running pass: remove-unused-names...                0.411736 seconds.
[PassRunner]   running pass: merge-blocks...                       0.896258 seconds.
[PassRunner]   running pass: precompute...                         1.24529 seconds.
[PassRunner]   running pass: optimize-instructions...              1.73888 seconds.
[PassRunner]   running pass: rse...                                3.88464 seconds.
[PassRunner]   running pass: vacuum...                             3.32334 seconds.
[PassRunner]   running pass: dae-optimizing...                     
Function Count: 132035

Iteration: 1
  removedParams: 12961
  returnValuesRemoved: 0
Iteration: 2
  removedParams: 2343
  returnValuesRemoved: 0
Iteration: 3
  removedParams: 903
  returnValuesRemoved: 0
Iteration: 4
  removedParams: 553
  returnValuesRemoved: 0
Iteration: 5
  removedParams: 409
  returnValuesRemoved: 0
Iteration: 6
  removedParams: 303
  returnValuesRemoved: 0
Iteration: 7
  removedParams: 124
  returnValuesRemoved: 0
Iteration: 8
  removedParams: 100
  returnValuesRemoved: 0
Iteration: 9
  removedParams: 88
  returnValuesRemoved: 0
Iteration: 10
  removedParams: 69
  returnValuesRemoved: 0
Iteration: 11
  removedParams: 53
  returnValuesRemoved: 0
Iteration: 12
  removedParams: 41
  returnValuesRemoved: 0
Iteration: 13
  removedParams: 21
  returnValuesRemoved: 0
Iteration: 14
  removedParams: 17
  returnValuesRemoved: 0
Iteration: 15
  removedParams: 12
  returnValuesRemoved: 0
Iteration: 16
  removedParams: 10
  returnValuesRemoved: 0
Iteration: 17
  removedParams: 7
  returnValuesRemoved: 0
Iteration: 18
  removedParams: 4
  returnValuesRemoved: 0
Iteration: 19
  removedParams: 4
  returnValuesRemoved: 0
Iteration: 20
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 21
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 22
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 23
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 24
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 25
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 26
  removedParams: 0
  returnValuesRemoved: 16824
Iteration: 27
  removedParams: 11
  returnValuesRemoved: 0
Iteration: 28
  removedParams: 1
  returnValuesRemoved: 0
Iteration: 29
  removedParams: 0
  returnValuesRemoved: 1523
Iteration: 30
  removedParams: 1
  returnValuesRemoved: 0
Iteration: 31
  removedParams: 0
  returnValuesRemoved: 101
Iteration: 32
  removedParams: 0
  returnValuesRemoved: 73
Iteration: 33
  removedParams: 0
  returnValuesRemoved: 24
Iteration: 34
  removedParams: 0
  returnValuesRemoved: 2
Iteration: 35
  removedParams: 0
  returnValuesRemoved: 1

Total Iterations: 35

Total removedParams: 18047

Total returnValuesRemoved : 18548
62.534 seconds.
[PassRunner]   running pass: inlining-optimizing...                19.0418 seconds.
[PassRunner]   running pass: duplicate-function-elimination...     1.20115 seconds.
[PassRunner]   running pass: duplicate-import-elimination...       0.0219464 seconds.
[PassRunner]   running pass: simplify-globals-optimizing...        0.238322 seconds.
[PassRunner]   running pass: remove-unused-module-elements...      0.744704 seconds.
[PassRunner]   running pass: directize...                          0.099373 seconds.
[PassRunner]   running pass: generate-stack-ir...                  3.8177 seconds.
[PassRunner]   running pass: optimize-stack-ir...                  11.7025 seconds.
[PassRunner]   running pass: strip-dwarf...                        0.381858 seconds.
[PassRunner]   running pass: post-emscripten...                    0.596744 seconds.
[PassRunner]   running pass: strip-debug...                        0.0331753 seconds.
[PassRunner]   running pass: strip-producers...                    0.0101664 seconds.
[PassRunner] passes took 206.203 seconds.

real	3m37.656s
user	8m52.359s
sys	0m1.559s

```

</details>
<details>
<summary>After</summary>

```
[PassRunner] running passes
[PassRunner]   running pass: duplicate-function-elimination...     15.0078 seconds.
[PassRunner]   running pass: memory-packing...                     0.390619 seconds.
[PassRunner]   running pass: once-reduction...                     0.32075 seconds.
[PassRunner]   running pass: ssa-nomerge...                        8.3342 seconds.
[PassRunner]   running pass: dce...                                2.6724 seconds.
[PassRunner]   running pass: remove-unused-names...                0.439206 seconds.
[PassRunner]   running pass: remove-unused-brs...                  2.43178 seconds.
[PassRunner]   running pass: remove-unused-names...                0.42836 seconds.
[PassRunner]   running pass: optimize-instructions...              2.01661 seconds.
[PassRunner]   running pass: pick-load-signs...                    0.60229 seconds.
[PassRunner]   running pass: precompute...                         1.29365 seconds.
[PassRunner]   running pass: optimize-added-constants-propagate... 15.0877 seconds.
[PassRunner]   running pass: code-pushing...                       0.825163 seconds.
[PassRunner]   running pass: simplify-locals-nostructure...        4.57273 seconds.
[PassRunner]   running pass: vacuum...                             3.6144 seconds.
[PassRunner]   running pass: reorder-locals...                     0.668802 seconds.
[PassRunner]   running pass: remove-unused-brs...                  1.7005 seconds.
[PassRunner]   running pass: coalesce-locals...                    3.7669 seconds.
[PassRunner]   running pass: local-cse...                          2.2485 seconds.
[PassRunner]   running pass: simplify-locals...                    5.48495 seconds.
[PassRunner]   running pass: vacuum...                             3.33565 seconds.
[PassRunner]   running pass: reorder-locals...                     0.575778 seconds.
[PassRunner]   running pass: coalesce-locals...                    3.47178 seconds.
[PassRunner]   running pass: reorder-locals...                     0.559923 seconds.
[PassRunner]   running pass: vacuum...                             3.30451 seconds.
[PassRunner]   running pass: code-folding...                       1.32623 seconds.
[PassRunner]   running pass: merge-blocks...                       0.969594 seconds.
[PassRunner]   running pass: remove-unused-brs...                  1.90931 seconds.
[PassRunner]   running pass: remove-unused-names...                0.401803 seconds.
[PassRunner]   running pass: merge-blocks...                       0.899873 seconds.
[PassRunner]   running pass: precompute...                         1.19759 seconds.
[PassRunner]   running pass: optimize-instructions...              1.73715 seconds.
[PassRunner]   running pass: rse...                                3.73981 seconds.
[PassRunner]   running pass: vacuum...                             3.36944 seconds.
[PassRunner]   running pass: dae-optimizing...                     
Function Count: 132035

Iteration: 1
  removedParams: 12961
  returnValuesRemoved: 15818
Iteration: 2
  removedParams: 2352
  returnValuesRemoved: 2472
Iteration: 3
  removedParams: 906
  returnValuesRemoved: 154
Iteration: 4
  removedParams: 553
  returnValuesRemoved: 78
Iteration: 5
  removedParams: 409
  returnValuesRemoved: 27
Iteration: 6
  removedParams: 303
  returnValuesRemoved: 2
Iteration: 7
  removedParams: 124
  returnValuesRemoved: 1
Iteration: 8
  removedParams: 100
  returnValuesRemoved: 0
Iteration: 9
  removedParams: 88
  returnValuesRemoved: 0
Iteration: 10
  removedParams: 69
  returnValuesRemoved: 0
Iteration: 11
  removedParams: 53
  returnValuesRemoved: 0
Iteration: 12
  removedParams: 41
  returnValuesRemoved: 0
Iteration: 13
  removedParams: 21
  returnValuesRemoved: 0
Iteration: 14
  removedParams: 17
  returnValuesRemoved: 0
Iteration: 15
  removedParams: 12
  returnValuesRemoved: 0
Iteration: 16
  removedParams: 10
  returnValuesRemoved: 0
Iteration: 17
  removedParams: 7
  returnValuesRemoved: 0
Iteration: 18
  removedParams: 4
  returnValuesRemoved: 0
Iteration: 19
  removedParams: 4
  returnValuesRemoved: 0
Iteration: 20
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 21
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 22
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 23
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 24
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 25
  removedParams: 2
  returnValuesRemoved: 0

Total Iterations: 25

Total removedParams: 18046

Total returnValuesRemoved : 18552
22.6605 seconds.
[PassRunner]   running pass: inlining-optimizing...                20.0691 seconds.
[PassRunner]   running pass: duplicate-function-elimination...     1.26333 seconds.
[PassRunner]   running pass: duplicate-import-elimination...       0.0230375 seconds.
[PassRunner]   running pass: simplify-globals-optimizing...        0.259489 seconds.
[PassRunner]   running pass: remove-unused-module-elements...      0.763686 seconds.
[PassRunner]   running pass: directize...                          0.0925571 seconds.
[PassRunner]   running pass: generate-stack-ir...                  0.653405 seconds.
[PassRunner]   running pass: optimize-stack-ir...                  12.3452 seconds.
[PassRunner]   running pass: strip-dwarf...                        0.419868 seconds.
[PassRunner]   running pass: post-emscripten...                    0.603769 seconds.
[PassRunner]   running pass: strip-debug...                        0.0351469 seconds.
[PassRunner]   running pass: strip-producers...                    0.0250297 seconds.
[PassRunner] passes took 157.92 seconds.

real	2m49.617s
user	7m59.884s
sys	0m22.072s
```

</details>

</details>