@kripken 2022-05-02
Nice improvement in those numbers!

Would it be possible to split the two main changes here into two separate PRs? After reading for a few minutes it's not obvious which changes are from which, so I think it would be easier to review separately. It would also be safer to land that way as if we later find a bug then bisection could pick between them.