@arsnyder16 2022-05-02
Yea not a problem, i'll look at splitting them up, the first bullet is fairly easy split,   might make sense for a 3rd PR as well with some more minor things